requirements.txt
opencv-python
numpy
mysql-connector-python

# run : pip3 install -r requirements.txt